Layout and testimonials update
