--- app/components/ui/hero-video-dialog-demo.tsx
@@ -1,50 +1,0 @@
-"use client"
-
-import { HeroVideoDialog } from "@/app/components/ui/hero-video-dialog"
-
-function HeroVideoDialogDemo() {
-  return (
-    <div className="relative">
-      <HeroVideoDialog
-        className="dark:hidden block"
-        animationStyle="from-center"
-        videoSrc="https://www.youtube.com/embed/qh3NGpYRG3I?si=4rb-zSdDkVK9qxxb"
-        thumbnailSrc="https://images.unsplash.com/photo-1518609878373-06d740f60d8b?q=80&w=1920&auto=format&fit=crop"
-        thumbnailAlt="Hero Video"
-      />
-      <HeroVideoDialog
-        className="hidden dark:block"
-        animationStyle="from-center"
-        videoSrc="https://www.youtube.com/embed/qh3NGpYRG3I?si=4rb-zSdDkVK9qxxb"
-        thumbnailSrc="https://images.unsplash.com/photo-1518609878373-06d740f60d8b?q=80&w=1920&auto=format&fit=crop"
-        thumbnailAlt="Hero Video"
-      />
-    </div>
-  )
-}
-
-function HeroVideoDialogDemoTopInBottomOut() {
-  return (
-    <div className="relative">
-      <HeroVideoDialog
-        className="dark:hidden block"
-        animationStyle="top-in-bottom-out"
-        videoSrc="https://www.youtube.com/embed/qh3NGpYRG3I?si=4rb-zSdDkVK9qxxb"
-        thumbnailSrc="https://images.unsplash.com/photo-1518609878373-06d740f60d8b?q=80&w=1920&auto=format&fit=crop"
-        thumbnailAlt="Hero Video"
-      />
-      <HeroVideoDialog
-        className="hidden dark:block"
-        animationStyle="top-in-bottom-out"
-        videoSrc="https://www.youtube.com/embed/qh3NGpYRG3I?si=4rb-zSdDkVK9qxxb"
-        thumbnailSrc="https://images.unsplash.com/photo-1518609878373-06d740f60d8b?q=80&w=1920&auto=format&fit=crop"
-        thumbnailAlt="Hero Video"
-      />
-    </div>
-  )
-}
-
-export {
-  HeroVideoDialogDemo,
-  HeroVideoDialogDemoTopInBottomOut,
-}
