+++ app/components/ui/hero-video.css
@@ -1,0 +1,12 @@
+.yt-embed-holder {
+  width: 100%;
+  overflow: hidden;
+  aspect-ratio: 16/9;
+  /* pointer-events: none; */
+}
+
+.yt-embed-holder iframe {
+  width: 300%;
+  height: 100%;
+  margin-left: -100%;
+}
